task_3/machine_learning/requirements.txt
# 기본 라이브러리
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.7.0

# 머신러닝 및 평가 지표
scikit-learn>=1.2.0


# PyTorch
# pip install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cu121

# Hugging Face Transformers (4.57.0)
# pip install git+https://github.com/huggingface/transformers
datasets

# 기타 유틸
tqdm>=4.65.0
sentencepiece 
protobuf 
tiktoken 
tokenizers
accelerate 
safetensors
pillow 
torchvision 
peft